feat: enhance delete feature

+++ src/main/java/id/ac/ui/cs/advprog/eshop/controller/ProductController.java
@@ -52,17 +52,12 @@ public class ProductController {
         return "redirect:/product/list";
     }
 
-    @PostMapping("delete/{id}")
-    public String deleteProduct(@PathVariable("id") String id, Model model) {
-        Product product = service.findById(id);
-        service.delete(product);
-        return "DeleteProduct";
-    }
-
     @PostMapping("/delete/{id}")
-    public String deleteProductPage(@PathVariable("id") String id, @ModelAttribute Product product) {
-        product.setProductId(id);
-        service.delete(product);
+    public String deleteProduct(@PathVariable("id") String id) {
+        Product product = service.findById(id);
+        if (product != null) {
+            service.delete(product);
+        }
         return "redirect:/product/list";
     }
 }
